+++ pkg/dataloaders/data_loader_middleware.go
@@ -1,4 +1,4 @@
-package loaders
+package dataloaders
 
 import (
 	"context"
