+++ pkg/dataloaders/user_account_loader.go
@@ -1,4 +1,4 @@
-package loaders
+package dataloaders
 
 import (
 	"context"
